@@ -1,6 +1,6 @@
 {
   "name": "@pact-foundation/pact-core",
-  "version": "10.12.1",
+  "version": "11.0.0",
   "description": "Core of @pact-foundation/pact. You almost certainly don't want to depend on this directly.",
   "main": "dist/src/index.js",
   "homepage": "https://github.com/pact-foundation/pact-js-core#readme",
@@ -51,7 +51,6 @@
   },
   "dependencies": {
     "@types/pino": "^6.3.5",
-    "@types/q": "1.0.7",
     "@types/request": "2.48.2",
     "chalk": "2.3.1",
     "check-types": "7.3.0",
@@ -62,7 +61,7 @@
     "node-gzip": "^1.1.2",
     "pino": "^6.11.0",
     "pino-pretty": "^4.1.0",
-    "q": "1.5.1",
+    "promise-timeout": "^1.3.0",
     "request": "2.88.0",
     "rimraf": "2.6.2",
     "sumchecker": "^2.0.2",
@@ -77,7 +76,7 @@
     "@types/basic-auth": "^1.1.2",
     "@types/chai": "4.1.2",
     "@types/chai-as-promised": "7.1.0",
-    "@types/check-types": "^7.3.1",
+    "@types/check-types": "^7.3.2",
     "@types/cors": "^2.8.6",
     "@types/cross-spawn": "^6.0.1",
     "@types/decompress": "^4.2.3",
@@ -88,9 +87,11 @@
     "@types/mocha": "2.2.48",
     "@types/node": "9.4.6",
     "@types/node-gzip": "^1.1.0",
+    "@types/promise-timeout": "^1.3.0",
     "@types/rimraf": "^2.0.2",
     "@types/tar": "^4.0.3",
     "@types/underscore": "1.8.7",
+    "@types/unixify": "^1.0.0",
     "@types/unzipper": "^0.10.2",
     "@types/url-join": "^4.0.0",
     "@typescript-eslint/eslint-plugin": "^2.3.2",
@@ -107,16 +108,16 @@
     "eslint-config-prettier": "^6.3.0",
     "eslint-plugin-prettier": "^3.1.1",
     "express": "4.16.2",
-    "jest": "^25.5.4",
+    "jest": "^26.5.6",
     "mocha": "^7.1.1",
     "nodemon": "^2.0.4",
     "prettier": "^1.18.2",
     "sinon": "9.2.4",
-    "snyk": "^1.230.5",
+    "snyk": "^1.605.0",
     "standard-version": "^9.1.0",
-    "ts-jest": "^25.5.1",
-    "ts-node": "8.3.0",
-    "typescript": "3.5.3"
+    "ts-jest": "^26.5.6",
+    "ts-node": "9.1.1",
+    "typescript": "4.2.4"
   },
   "scripts": {
     "postinstall": "node postinstall.js",
