@@ -1,20 +1,13 @@
-import q = require('q');
 import path = require('path');
 import fs = require('fs');
 import logger, { verboseIsImplied } from './logger';
 import spawn from './spawn';
+import { timeout, TimeoutError } from 'promise-timeout';
 import { DEFAULT_ARG } from './spawn';
-import { deprecate } from 'util';
 import pactStandalone from './pact-standalone';
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const checkTypes = require('check-types');
+import checkTypes = require('check-types');
 
 export class Publisher {
-  public static create = deprecate(
-    (options: PublisherOptions) => new Publisher(options),
-    'Create function will be removed in future release, please use the default export function or use `new Publisher()`'
-  );
-
   public readonly options: PublisherOptions;
   private readonly __argMapping = {
     pactFilesOrDirs: DEFAULT_ARG,
@@ -102,33 +95,38 @@ export class Publisher {
     this.options = options;
   }
 
-  public publish(): q.Promise<string[]> {
+  public publish(): Promise<string[]> {
     logger.info(`Publishing pacts to broker at: ${this.options.pactBroker}`);
-    const deferred = q.defer<string[]>();
-    const instance = spawn.spawnBinary(
-      pactStandalone.brokerPath,
-      [{ cliVerb: 'publish' }, this.options],
-      this.__argMapping
-    );
-    const output: Array<string | Buffer> = [];
-    instance.stdout.on('data', l => output.push(l));
-    instance.stderr.on('data', l => output.push(l));
-    instance.once('close', code => {
-      const o = output.join('\n');
-      const pactUrls = /^https?:\/\/.*\/pacts\/.*$/gim.exec(o);
-      if (code !== 0 || !pactUrls) {
-        logger.error(`Could not publish pact:\n${o}`);
-        return deferred.reject(new Error(o));
+
+    return timeout(
+      new Promise<string[]>((resolve, reject) => {
+        const instance = spawn.spawnBinary(
+          pactStandalone.brokerPath,
+          [{ cliVerb: 'publish' }, this.options],
+          this.__argMapping
+        );
+        const output: Array<string | Buffer> = [];
+        instance.stdout.on('data', l => output.push(l));
+        instance.stderr.on('data', l => output.push(l));
+        instance.once('close', code => {
+          const o = output.join('\n');
+          const pactUrls = /^https?:\/\/.*\/pacts\/.*$/gim.exec(o);
+          if (code !== 0 || !pactUrls) {
+            logger.error(`Could not publish pact:\n${o}`);
+            return reject(new Error(o));
+          }
+
+          logger.info(o);
+          return resolve(pactUrls);
+        });
+      }),
+      this.options.timeout as number
+    ).catch((err: Error) => {
+      if (err instanceof TimeoutError) {
+        throw new Error(`Timeout waiting for publication process to complete`);
       }
-
-      logger.info(o);
-      return deferred.resolve(pactUrls);
+      throw err;
     });
-
-    return deferred.promise.timeout(
-      this.options.timeout as number,
-      `Timeout waiting for verification process to complete (PID: ${instance.pid})`
-    );
   }
 }
 
