@@ -1,19 +1,11 @@
 import { AbstractService, LogLevel } from './service';
-import { deprecate } from 'util';
 import pact from './pact-standalone';
 import path = require('path');
 import fs = require('fs');
 import mkdirp = require('mkdirp');
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const checkTypes = require('check-types');
+import checkTypes = require('check-types');
 
 export class Server extends AbstractService {
-  public static create = deprecate(
-    (options?: ServerOptions) => new Server(options),
-    'Create function will be removed in future release, please use the default export function or use `new Server()`'
-  );
-
   public readonly options: ServerOptions;
 
   constructor(options: ServerOptions = {}) {
