@@ -1,17 +1,10 @@
 import { DEFAULT_ARG } from './spawn';
 import { AbstractService, LogLevel } from './service';
-import { deprecate } from 'util';
 
 import pact from './pact-standalone';
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const checkTypes = require('check-types');
+import checkTypes = require('check-types');
 
 export class Stub extends AbstractService {
-  public static create = deprecate(
-    (options?: StubOptions) => new Stub(options),
-    'Create function will be removed in future release, please use the default export function or use `new Stub()`'
-  );
-
   public readonly options: StubOptions;
 
   constructor(options?: StubOptions) {
