@@ -1,4 +1,3 @@
-import q = require('q');
 import express = require('express');
 import * as http from 'http';
 import { auth, returnJson } from './data-utils';
@@ -6,7 +5,7 @@ import cors = require('cors');
 import _ = require('underscore');
 import bodyParser = require('body-parser');
 
-export default (port: number): q.Promise<http.Server> => {
+export default (port: number): Promise<http.Server> => {
   const BROKER_HOST = `http://localhost:${port}`;
   const server: express.Express = express();
   server.use(cors());
@@ -531,7 +530,8 @@ export default (port: number): q.Promise<http.Server> => {
     })
   );
 
-  const deferred = q.defer<http.Server>();
-  let s = server.listen(port, () => deferred.resolve());
-  return deferred.promise.then(() => s);
+  let s: http.Server;
+  return new Promise<void>(resolve => {
+    s = server.listen(port, () => resolve());
+  }).then(() => s);
 };
